+++ ItemQualities/Assets/ItemQualities/Scripts/ItemQualitiesContent.cs
@@ -822,6 +822,8 @@ namespace ItemQualities
             public static BuffQualityGroup Warbanner;
 
             public static BuffQualityGroup DelayedDamageDebuff;
+
+            public static BuffQualityGroup CrowbarCharge;
         }
 
         public static class Buffs